backend/requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
aiofiles
python-multipart
python-dotenv
pymysql